requirements.txt
alpaca-py==3.3.0
mcp==1.0.0
python-dotenv==1.0.1
pydantic==2.7.4